decupled MasterList
